Split out all rendering into and damage handling into rendering.c
